pyproject.toml
[project]
name = "pycon-apac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5.20",
    "langchain-community>=0.3.8",
    "langchain>=0.3.9",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "sentence-transformers>=3.3.1",
    "langchain-huggingface>=0.1.2",
    "sentencepiece>=0.2.0",
    "langchain-openai>=0.2.10",
    "python-dotenv>=1.0.1",
    "lark>=1.2.2",
    "structlog>=24.4.0",
    "openai>=1.55.3",
    "anthropic>=0.40.0",
    "requests>=2.32.3",
]
